added logout button component
